\docType{data}
\name{iris}
\alias{iris}
\title{An example data frame Iris}
\description{
  An example data frame with iris data \itemize{
  \item{Sepal.Width:}{Sepal width}
  \item{Sepal.Length:}{Sepal Length}
  \item{Petal.Width:}{Petal Width}
  \item{Petal.Length:}{Petal Length} \item{Species:}{Sales
  value} }
}
\examples{
data(iris)
str(iris)
}
\keyword{data}
\keyword{datasets}

