\name{print.myclassname}
\alias{print.myclassname}
\title{Mytitle}
\usage{
  \method{print}{myclassname} (x, ...)
}
\arguments{
  \item{x}{an object of class myclassname}

  \item{...}{further arguments passed to or from other
  methods.}
}
\value{
  Nothing, prints
}
\description{
  My description
}
\examples{
data(iris)
myfunction(iris)
}
\seealso{
  \code{\link{print}}
}

