\name{readXLSModelInput}
\alias{readXLSModelInput}
\title{read accent model input from an XLS file}
\usage{
  readXLSModelInput(xlsFile)
}
\arguments{
  \item{xlsFile}{\itemize{ \item{sheet patients:}{Sepal
  width} \item{sheet.therapist:}{Sepal Length}
  \item{sheet.links:}{Petal Width} }}
}
\value{
  AccentModelInput instance. This instance contains 3
  data.frames extracted from the given XLS file.
}
\description{
  desc
}

