\docType{package}
\name{thinkdata.accent}
\alias{thinkdata.accent}
\alias{thinkdata.accent-package}
\title{Package title here}
\description{
  Package description here
}
\section{Data}{
  \tabular{ll}{ \code{\link{iris}:} \tab An example data
  frame \cr }
}

\section{Utilities}{
  \tabular{ll}{ \code{\link{myfunction}:} \tab TODO. \cr }
}
\examples{
# See the specified functions in the package
}
\author{
  Christophe Van Huele
  \email{christophe.vanhuele@gmail.com}

  Kenny Helsens \email{kenny@thinkdata.be}
}

